Represented attributes are not included by default
