+++ lib/active_data/model/associations.rb
@@ -28,7 +28,6 @@ module ActiveData
 
       included do
         include NestedAttributes
-        include Validations
 
         class_attribute :_associations, :_association_aliases, instance_reader: false, instance_writer: false
         self._associations = {}
