+++ lib/active_data/model/associations/validations.rb
@@ -19,12 +19,6 @@ module ActiveData
 
       private
 
-        def run_validations! #:nodoc:
-          super
-          emerge_represented_attributes_errors!
-          errors.empty?
-        end
-
         def validate_nested!
           association_names.each do |name|
             association = association(name)
@@ -36,26 +30,6 @@ module ActiveData
 
             ActiveData::Model::Validations::NestedValidator
               .validate_nested(self, name, association.target, &invalid_block)
-          end
-        end
-
-        # Move represent attribute errors to the top level:
-        #
-        #   {:'role.email' => ['Some error']}
-        #
-        # to:
-        #
-        #   {email: ['Some error']}
-        #
-        def emerge_represented_attributes_errors!
-          self.class.represented_attributes.each do |attribute|
-            key = :"#{attribute.reference}.#{attribute.column}"
-            # Rails 5 pollutes messages with an empty array on key data fetch attempt
-            messages = errors.messages[key] if errors.messages.key?(key)
-            if messages.present?
-              errors[attribute.column].concat(messages)
-              errors.delete(key)
-            end
           end
         end
       end
