--- lib/active_data/persistence_adapters.rb
@@ -1,2 +1,0 @@
-require 'active_data/model/associations/persistence_adapters/base'
-require 'active_data/model/associations/persistence_adapters/active_record'
