+++ spec/lib/active_data/model/attributes/represents_spec.rb
@@ -83,78 +83,4 @@ describe ActiveData::Model::Attributes::Represents do
       specify { expect { sleep(0.01) }.not_to change { field.read_before_type_cast } }
     end
   end
-
-  context 'integration' do
-    before do
-      stub_model(:author) do
-        attribute :rate, Integer
-      end
-
-      stub_model(:post) do
-        attribute :author, Object
-        alias_attribute :a, :author
-        represents :rate, of: :a
-        alias_attribute :r, :rate
-      end
-    end
-    let(:author) { Author.new(rate: '42') }
-
-    specify { expect(Post.reflect_on_attribute(:rate).reference).to eq('author') }
-
-    specify { expect(Post.new(author: author).rate).to eq(42) }
-    specify { expect(Post.new(author: author).rate_before_type_cast).to eq('42') }
-    specify { expect(Post.new(rate: '33', author: author).rate).to eq(33) }
-    specify { expect(Post.new(rate: '33', author: author).author.rate).to eq(33) }
-    specify { expect(Post.new(r: '33', author: author).rate).to eq(33) }
-    specify { expect(Post.new(r: '33', author: author).author.rate).to eq(33) }
-    specify { expect(Post.new(author: author).rate?).to eq(true) }
-    specify { expect(Post.new(rate: nil, author: author).rate?).to eq(false) }
-
-    specify { expect(Post.new.rate).to be_nil }
-    specify { expect(Post.new.rate_before_type_cast).to be_nil }
-    specify { expect { Post.new(rate: '33') }.to raise_error(NoMethodError) }
-
-    context do
-      before do
-        stub_class(:author, ActiveRecord::Base)
-
-        stub_model(:post) do
-          include ActiveData::Model::Associations
-
-          references_one :author
-          alias_association :a, :author
-          represents :name, of: :a
-        end
-      end
-      let!(:author) { Author.create!(name: 42) }
-
-      specify { expect(Post.reflect_on_attribute(:name).reference).to eq('author') }
-
-      specify { expect(Post.new(name: '33', author: author).name).to eq('33') }
-      specify { expect(Post.new(name: '33', author: author).author.name).to eq('33') }
-    end
-
-    context 'multiple attributes in a single represents definition' do
-      before do
-        stub_model(:author) do
-          attribute :first_name, String
-          attribute :last_name, String
-        end
-
-        stub_model(:post) do
-          attribute :author, Object
-          represents :first_name, :last_name, of: :author
-        end
-      end
-
-      let(:author) { Author.new(first_name: 'John', last_name: 'Doe') }
-      let(:post) { Post.new }
-
-      specify do
-        expect { post.update(author: author) }
-          .to change { post.first_name }.to('John')
-          .and change { post.last_name }.to('Doe')
-      end
-    end
-  end
 end
